+++ src/app/_models/query.ts
@@ -1,0 +1,10 @@
+import {DataSource} from './data.source';
+
+export class Query {
+    id: number;
+    name: string;
+    query: string;
+    user_id: number;
+    data_source_id: number;
+    data_source: DataSource;
+}